refactor: 초대 관련 타입 분리 및 add_edit_member RPC p_resource_id 파라미터 추가
