+++ components/features/invite/InviteEditorHandler.tsx
@@ -1,9 +1,10 @@
 'use client';
 
 import { useInviteEditorHandler } from '@/hooks/invite-member/useInviteEditorHandler';
-import { ResourceType } from '@/lib/actions/invite';
+import { ResourceType } from '@/types/invite';
 
-export default function InviteEditorHandler({ type }: { type: ResourceType }) {
-  useInviteEditorHandler(type);
+// router를 사용하는 훅을 실행시키기 위한 컴포넌트로, 아무것도 반환하지 X
+export default function InviteEditorHandler({ id, type }: { id: string; type: ResourceType }) {
+  useInviteEditorHandler(id, type);
   return null;
 }
