+++ hooks/invite-member/useAddEditMember.ts
@@ -1,8 +1,9 @@
-import { addEditMember, ResourceType } from '@/lib/actions/invite';
+import { addEditMember } from '@/lib/actions/invite';
+import { AddEditMemberParams } from '@/types/invite';
 import { useMutation } from '@tanstack/react-query';
 
 export const useAddEditMember = () => {
   return useMutation({
-    mutationFn: ({ token, type }: { token: string; type: ResourceType }) => addEditMember(token, type),
+    mutationFn: (params: AddEditMemberParams) => addEditMember(params),
   });
 };
